Change the name of functions and their variables
I changed the name of the functions that generate
tiles to match the task they were doing better. I
also, changed the name of the variables inside the
same functions so that they would remain
associated with the task that their respective
function is carrying out.

Note: Elaborating on my previous commit, the
reason why I made more functions to contain code
was to make my code more concise and readable.